Moved tests to the new view model
