+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/mystore/MyStoreViewModelTest.kt
@@ -2,23 +2,16 @@ package com.woocommerce.android.ui.mystore
 
 import androidx.lifecycle.SavedStateHandle
 import com.woocommerce.android.AppPrefsWrapper
-import com.woocommerce.android.AppUrls
-import com.woocommerce.android.BuildConfig
 import com.woocommerce.android.WooException
 import com.woocommerce.android.analytics.AnalyticsEvent
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
-import com.woocommerce.android.model.UiString
 import com.woocommerce.android.tools.NetworkStatus
 import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.tools.SiteConnectionType
-import com.woocommerce.android.ui.jitm.JitmTracker
-import com.woocommerce.android.ui.jitm.QueryParamsEncoder
-import com.woocommerce.android.ui.mystore.MyStoreViewModel.Companion.UTM_SOURCE
 import com.woocommerce.android.ui.mystore.domain.GetStats
 import com.woocommerce.android.ui.mystore.domain.GetTopPerformers
 import com.woocommerce.android.ui.mystore.domain.GetTopPerformers.TopPerformerProduct
-import com.woocommerce.android.ui.payments.banner.BannerState
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import com.woocommerce.android.viewmodel.ResourceProvider
@@ -30,24 +23,15 @@ import org.junit.Before
 import org.junit.Test
 import org.mockito.ArgumentMatchers.anyBoolean
 import org.mockito.ArgumentMatchers.anyInt
-import org.mockito.ArgumentMatchers.anyString
 import org.mockito.kotlin.any
-import org.mockito.kotlin.argumentCaptor
-import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.never
-import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
 import org.wordpress.android.fluxc.model.SiteModel
 import org.wordpress.android.fluxc.network.BaseRequest
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooError
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooErrorType
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooResult
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.jitm.JITMApiResponse
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.jitm.JITMContent
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.jitm.JITMCta
-import org.wordpress.android.fluxc.store.JitmStore
 import org.wordpress.android.fluxc.store.WCStatsStore.StatsGranularity
 import org.wordpress.android.fluxc.store.WooCommerceStore
 import kotlin.test.assertTrue
@@ -65,10 +49,6 @@ class MyStoreViewModelTest : BaseUnitTest() {
     private val appPrefsWrapper: AppPrefsWrapper = mock()
     private val usageTracksEventEmitter: MyStoreStatsUsageTracksEventEmitter = mock()
     private val analyticsTrackerWrapper: AnalyticsTrackerWrapper = mock()
-    private val jitmStore: JitmStore = mock()
-    private val jitmTracker: JitmTracker = mock()
-    private val utmProvider: MyStoreUtmProvider = mock()
-    private val queryParamsEncoder: QueryParamsEncoder = mock()
 
     private lateinit var sut: MyStoreViewModel
 
@@ -76,9 +56,6 @@ class MyStoreViewModelTest : BaseUnitTest() {
     fun setup() = testBlocking {
         givenStatsLoadingResult(GetStats.LoadStatsResult.VisitorsStatsError)
         givenFetchTopPerformersResult(Result.failure(WooException(WOO_GENERIC_ERROR)))
-        whenever(queryParamsEncoder.getEncodedQueryParams()).thenReturn(
-            "build_type=developer&platform=android&version=${BuildConfig.VERSION_NAME}"
-        )
     }
 
     @Test
@@ -435,844 +412,6 @@ class MyStoreViewModelTest : BaseUnitTest() {
             )
         }
 
-    // region Just In Time Messages (JITM)
-    @Test
-    fun `given network connection, when on swipe to refresh, then fetch JITMS`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenViewModelIsCreated()
-
-            sut.onPullToRefresh()
-
-            // called twice, on view model init and on pull to refresh
-            verify(jitmStore, times(2)).fetchJitmMessage(any(), any(), any())
-        }
-    }
-
-    @Test
-    fun `given store setup in US, when viewmodel init, then request for jitm with valid message path`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            val expectedMessagePath = "woomobile:my_store:admin_notices"
-            val captor = argumentCaptor<String>()
-
-            whenViewModelIsCreated()
-            verify(jitmStore).fetchJitmMessage(any(), captor.capture(), any())
-
-            assertThat(captor.firstValue).isEqualTo(expectedMessagePath)
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then proper banner state event is triggered`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(sut.bannerState.value).isInstanceOf(BannerState::class.java)
-        }
-    }
-
-    @Test
-    fun `given jitm error response, when viewmodel init, then banner state event is not triggered`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    error = WOO_GENERIC_ERROR
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(sut.bannerState.value).isNull()
-        }
-    }
-
-    @Test
-    fun `given jitm empty response, when viewmodel init, then banner state hide event is triggered`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = emptyArray()
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(sut.bannerState.value).isInstanceOf(BannerState.HideBannerState::class.java)
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then proper jitm message is used in UI`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            val testJitmMessage = "Test jitm message"
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            content = provideJitmContent(message = testJitmMessage)
-                        )
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(
-                (
-                    (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                    ).title
-            ).isEqualTo(
-                UiString.UiStringText(text = testJitmMessage, containsHtml = false)
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then proper jitm description is used in UI`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            val testJitmDescription = "Test jitm description"
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            content = provideJitmContent(description = testJitmDescription)
-                        )
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(
-                (
-                    (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                    ).description
-            ).isEqualTo(
-                UiString.UiStringText(text = testJitmDescription, containsHtml = false)
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then proper jitm cta label is used in UI`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            val testJitmCtaLabel = "Test jitm Cta label"
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            jitmCta = provideJitmCta(message = testJitmCtaLabel)
-                        )
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            assertThat(
-                (
-                    (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                    ).primaryActionLabel
-            ).isEqualTo(
-                UiString.UiStringText(text = testJitmCtaLabel, containsHtml = false)
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when jitm cta clicked, then jitm click event emitted`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(selectedSite.getIfExists()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-            whenever(
-                utmProvider.getUrlWithUtmParams(
-                    anyString(),
-                    anyString(),
-                    anyString(),
-                    any(),
-                    anyString(),
-                )
-            ).thenReturn("")
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onPrimaryActionClicked.invoke()
-
-            assertThat(sut.event.value).isInstanceOf(MyStoreViewModel.MyStoreEvent.OpenJITMAction::class.java)
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when jitm cta clicked, then proper url is passedto OpenJITM event`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(selectedSite.getIfExists()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            jitmCta = provideJitmCta(
-                                link = "${AppUrls.WOOCOMMERCE_PURCHASE_CARD_READER_IN_COUNTRY}US"
-                            )
-                        )
-                    )
-                )
-            )
-            whenever(
-                utmProvider.getUrlWithUtmParams(
-                    anyString(),
-                    anyString(),
-                    anyString(),
-                    any(),
-                    anyString(),
-                )
-            ).thenReturn(
-                "${AppUrls.WOOCOMMERCE_PURCHASE_CARD_READER_IN_COUNTRY}US"
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onPrimaryActionClicked.invoke()
-
-            assertThat(sut.event.value as MyStoreViewModel.MyStoreEvent.OpenJITMAction).isEqualTo(
-                MyStoreViewModel.MyStoreEvent.OpenJITMAction(
-                    "${AppUrls.WOOCOMMERCE_PURCHASE_CARD_READER_IN_COUNTRY}US"
-                )
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when jitm dismiss tapped, then banner state is updated to not display`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            assertThat(sut.bannerState.value).isInstanceOf(BannerState.HideBannerState::class.java)
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then jitm fetch success is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchSuccess(
-                any(),
-                any(),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success, when viewmodel init, then jitm fetch success is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse(id = "12345"))
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchSuccess(
-                UTM_SOURCE,
-                "12345",
-                1
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success, when viewmodel init, then jitm fetch success is tracked with highest score jitm id`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(id = "12345"),
-                        provideJitmApiResponse(id = "123456"),
-                        provideJitmApiResponse(id = "123")
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchSuccess(
-                UTM_SOURCE,
-                "12345",
-                3
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success response, when viewmodel init, then jitm displayed is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmDisplayed(
-                any(),
-                any(),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success, when viewmodel init, then jitm displayed is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmDisplayed(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success with empty jitms, when viewmodel init, then jitm fetch success is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = emptyArray()
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchSuccess(
-                anyString(),
-                eq(null),
-                anyInt()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm success with empty jitms, when viewmodel init, then event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = emptyArray()
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchSuccess(
-                UTM_SOURCE,
-                null,
-                0
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm failure response, when viewmodel init, then jitm fetch failure is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    WooError(
-                        type = WooErrorType.GENERIC_ERROR,
-                        original = BaseRequest.GenericErrorType.NETWORK_ERROR,
-                        message = ""
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchFailure(anyString(), any(), anyString())
-        }
-    }
-
-    @Test
-    fun `given jitm failure, when viewmodel init, then jitm fetch failure is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    WooError(
-                        type = WooErrorType.GENERIC_ERROR,
-                        original = BaseRequest.GenericErrorType.NETWORK_ERROR,
-                        message = "Generic error"
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-
-            verify(jitmTracker).trackJitmFetchFailure(
-                UTM_SOURCE,
-                WooErrorType.GENERIC_ERROR,
-                "Generic error"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when cta tapped, then cta tapped event is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(selectedSite.getIfExists()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-            whenever(
-                utmProvider.getUrlWithUtmParams(
-                    anyString(),
-                    anyString(),
-                    anyString(),
-                    any(),
-                    anyString(),
-                )
-            ).thenReturn("")
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onPrimaryActionClicked.invoke()
-
-            verify(jitmTracker).trackJitmCtaTapped(
-                any(),
-                any(),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when cta tapped, then cta tapped event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(selectedSite.getIfExists()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-            whenever(
-                utmProvider.getUrlWithUtmParams(
-                    anyString(),
-                    anyString(),
-                    anyString(),
-                    any(),
-                    anyString(),
-                )
-            ).thenReturn("")
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onPrimaryActionClicked.invoke()
-
-            verify(jitmTracker).trackJitmCtaTapped(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when dismiss tapped, then dismiss tapped event is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissTapped(
-                any(),
-                any(),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm displayed, when dismiss tapped, then dismiss tapped event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissTapped(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss success, then dismiss success event is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(true)
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissSuccess(
-                any(),
-                any(),
-                any()
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss success, then dismiss success event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(true)
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissSuccess(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss failure, then dismiss failure event is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(false)
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissFailure(
-                anyString(),
-                anyString(),
-                anyString(),
-                eq(null),
-                eq(null)
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss error, then dismiss failure event is tracked`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(provideJitmApiResponse())
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(
-                    WooError(
-                        type = WooErrorType.GENERIC_ERROR,
-                        original = BaseRequest.GenericErrorType.NETWORK_ERROR
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissFailure(
-                anyString(),
-                anyString(),
-                anyString(),
-                any(),
-                eq(null)
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss error, then dismiss failure event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(
-                    WooError(
-                        type = WooErrorType.GENERIC_ERROR,
-                        original = BaseRequest.GenericErrorType.NETWORK_ERROR,
-                        message = "Generic error"
-                    )
-                )
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissFailure(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp",
-                WooErrorType.GENERIC_ERROR,
-                "Generic error"
-            )
-        }
-    }
-
-    @Test
-    fun `given jitm dismissed, when dismiss failure, then dismiss failure event is tracked with correct properties`() {
-        testBlocking {
-            givenNetworkConnectivity(connected = true)
-            whenever(selectedSite.get()).thenReturn(SiteModel())
-            whenever(
-                jitmStore.fetchJitmMessage(any(), any(), any())
-            ).thenReturn(
-                WooResult(
-                    model = arrayOf(
-                        provideJitmApiResponse(
-                            id = "12345",
-                            featureClass = "woomobile_ipp"
-                        )
-                    )
-                )
-            )
-            whenever(jitmStore.dismissJitmMessage(any(), any(), any())).thenReturn(
-                WooResult(false)
-            )
-
-            whenViewModelIsCreated()
-            (
-                (sut.bannerState.value as BannerState) as BannerState.DisplayBannerState
-                ).onDismissClicked.invoke()
-
-            verify(jitmTracker).trackJitmDismissFailure(
-                UTM_SOURCE,
-                "12345",
-                "woomobile_ipp",
-                null,
-                null
-            )
-        }
-    }
-    //endregion
-
     private suspend fun givenStatsLoadingResult(result: GetStats.LoadStatsResult) {
         whenever(getStats.invoke(any(), any())).thenReturn(flow { emit(result) })
     }
@@ -1326,10 +465,6 @@ class MyStoreViewModelTest : BaseUnitTest() {
             usageTracksEventEmitter,
             analyticsTrackerWrapper,
             myStoreTransactionLauncher = mock(),
-            jitmStore,
-            jitmTracker,
-            utmProvider,
-            queryParamsEncoder,
         )
     }
 
@@ -1351,48 +486,4 @@ class MyStoreViewModelTest : BaseUnitTest() {
             imageUrl = null
         )
     }
-
-    private fun provideJitmApiResponse(
-        content: JITMContent = provideJitmContent(),
-        jitmCta: JITMCta = provideJitmCta(),
-        timeToLive: Int = 0,
-        id: String = "",
-        featureClass: String = "",
-        expires: Long = 0L,
-        maxDismissal: Int = 2,
-        isDismissible: Boolean = false,
-        url: String = "",
-        jitmStatsUrl: String = ""
-    ) = JITMApiResponse(
-        content = content,
-        cta = jitmCta,
-        timeToLive = timeToLive,
-        id = id,
-        featureClass = featureClass,
-        expires = expires,
-        maxDismissal = maxDismissal,
-        isDismissible = isDismissible,
-        url = url,
-        jitmStatsUrl = jitmStatsUrl
-    )
-
-    private fun provideJitmContent(
-        message: String = "",
-        description: String = "",
-        icon: String = "",
-        title: String = ""
-    ) = JITMContent(
-        message = message,
-        description = description,
-        icon = icon,
-        title = title
-    )
-
-    private fun provideJitmCta(
-        message: String = "",
-        link: String = ""
-    ) = JITMCta(
-        message = message,
-        link = link
-    )
 }
